feat(cart): implement cart feature (data layer)
